pyproject.toml
[project]
name = "college-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.1",
    "datasets==3.6.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "gradio>=6.4.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter-dash>=0.4.2",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "modal>=1.3.1",
    "numpy>=2.4.1",
    "ollama>=0.6.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "protobuf==3.20.2",
    "psutil>=7.2.1",
    "pydub>=0.25.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "sentence-transformers>=5.2.0",
    "setuptools>=80.10.2",
    "speedtest-cli>=2.1.3",
    "torch>=2.10.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.6",
    "wandb>=0.24.0",
]
